Temporarily move API to "beta" package
