+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/HolographicDisplaysAPI.java
@@ -3,17 +3,17 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api;
+package me.filoghost.holographicdisplays.api.beta;
 
-import me.filoghost.holographicdisplays.api.hologram.Hologram;
-import me.filoghost.holographicdisplays.api.internal.HolographicDisplaysAPIProvider;
-import me.filoghost.holographicdisplays.api.placeholder.GlobalPlaceholder;
-import me.filoghost.holographicdisplays.api.placeholder.GlobalPlaceholderFactory;
-import me.filoghost.holographicdisplays.api.placeholder.GlobalPlaceholderReplacementSupplier;
-import me.filoghost.holographicdisplays.api.placeholder.IndividualPlaceholder;
-import me.filoghost.holographicdisplays.api.placeholder.IndividualPlaceholderFactory;
-import me.filoghost.holographicdisplays.api.placeholder.IndividualPlaceholderReplacementSupplier;
-import me.filoghost.holographicdisplays.api.placeholder.RegisteredPlaceholder;
+import me.filoghost.holographicdisplays.api.beta.hologram.Hologram;
+import me.filoghost.holographicdisplays.api.beta.internal.HolographicDisplaysAPIProvider;
+import me.filoghost.holographicdisplays.api.beta.placeholder.GlobalPlaceholder;
+import me.filoghost.holographicdisplays.api.beta.placeholder.GlobalPlaceholderFactory;
+import me.filoghost.holographicdisplays.api.beta.placeholder.GlobalPlaceholderReplacementSupplier;
+import me.filoghost.holographicdisplays.api.beta.placeholder.IndividualPlaceholder;
+import me.filoghost.holographicdisplays.api.beta.placeholder.IndividualPlaceholderFactory;
+import me.filoghost.holographicdisplays.api.beta.placeholder.IndividualPlaceholderReplacementSupplier;
+import me.filoghost.holographicdisplays.api.beta.placeholder.RegisteredPlaceholder;
 import org.bukkit.Location;
 import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.NotNull;
