+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/hologram/VisibilitySettings.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.hologram;
+package me.filoghost.holographicdisplays.api.beta.hologram;
 
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
