+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/hologram/line/HologramLinePickupEvent.java
@@ -3,11 +3,11 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.hologram.line;
+package me.filoghost.holographicdisplays.api.beta.hologram.line;
 
 import org.bukkit.entity.Player;
 
-public interface HologramLineClickEvent {
+public interface HologramLinePickupEvent {
 
     Player getPlayer();
