+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/hologram/line/ItemHologramLine.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.hologram.line;
+package me.filoghost.holographicdisplays.api.beta.hologram.line;
 
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
