+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/placeholder/GlobalPlaceholderFactory.java
@@ -3,18 +3,18 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.placeholder;
+package me.filoghost.holographicdisplays.api.beta.placeholder;
 
 import org.jetbrains.annotations.Nullable;
 
 /**
  * @since 1
  */
-public interface IndividualPlaceholderFactory {
+public interface GlobalPlaceholderFactory {
 
     /**
      * @since 1
      */
-    @Nullable IndividualPlaceholder getPlaceholder(@Nullable String argument);
+    @Nullable GlobalPlaceholder getPlaceholder(@Nullable String argument);
 
 }
