+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/placeholder/IndividualPlaceholderFactory.java
@@ -3,18 +3,18 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.placeholder;
+package me.filoghost.holographicdisplays.api.beta.placeholder;
 
 import org.jetbrains.annotations.Nullable;
 
 /**
  * @since 1
  */
-public interface GlobalPlaceholderFactory {
+public interface IndividualPlaceholderFactory {
 
     /**
      * @since 1
      */
-    @Nullable GlobalPlaceholder getPlaceholder(@Nullable String argument);
+    @Nullable IndividualPlaceholder getPlaceholder(@Nullable String argument);
 
 }
