+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/placeholder/IndividualPlaceholderReplacementSupplier.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.placeholder;
+package me.filoghost.holographicdisplays.api.beta.placeholder;
 
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
@@ -12,9 +12,12 @@ import org.jetbrains.annotations.Nullable;
 /**
  * @since 1
  */
-public interface IndividualPlaceholder extends Placeholder {
+@FunctionalInterface
+public interface IndividualPlaceholderReplacementSupplier {
 
     /**
+     * @see IndividualPlaceholder#getReplacement(Player, String)
+     *
      * @since 1
      */
     @Nullable String getReplacement(@NotNull Player player, @Nullable String argument);
