+++ api/src/main/java/me/filoghost/holographicdisplays/api/beta/placeholder/RegisteredPlaceholder.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package me.filoghost.holographicdisplays.api.placeholder;
+package me.filoghost.holographicdisplays.api.beta.placeholder;
 
 import org.jetbrains.annotations.NotNull;
