+++ legacy-api/v2/src/main/java/com/gmail/filoghost/holographicdisplays/api/internal/HologramsAPIProvider.java
@@ -7,7 +7,7 @@ package com.gmail.filoghost.holographicdisplays.api.internal;
 
 import com.gmail.filoghost.holographicdisplays.api.Hologram;
 import com.gmail.filoghost.holographicdisplays.api.placeholder.PlaceholderReplacer;
-import me.filoghost.holographicdisplays.api.internal.HolographicDisplaysAPIProvider;
+import me.filoghost.holographicdisplays.api.beta.internal.HolographicDisplaysAPIProvider;
 import org.bukkit.Location;
 import org.bukkit.entity.Entity;
 import org.bukkit.plugin.Plugin;
