+++ plugin/src/main/java/me/filoghost/holographicdisplays/plugin/api/current/APIHologramLine.java
@@ -5,7 +5,7 @@
  */
 package me.filoghost.holographicdisplays.plugin.api.current;
 
-import me.filoghost.holographicdisplays.api.hologram.line.HologramLine;
+import me.filoghost.holographicdisplays.api.beta.hologram.line.HologramLine;
 import me.filoghost.holographicdisplays.plugin.hologram.base.EditableHologramLine;
 import org.bukkit.plugin.Plugin;
