+++ plugin/src/main/java/me/filoghost/holographicdisplays/plugin/api/current/SimpleHologramLineClickEvent.java
@@ -5,7 +5,7 @@
  */
 package me.filoghost.holographicdisplays.plugin.api.current;
 
-import me.filoghost.holographicdisplays.api.hologram.line.HologramLineClickEvent;
+import me.filoghost.holographicdisplays.api.beta.hologram.line.HologramLineClickEvent;
 import org.bukkit.entity.Player;
 
 class SimpleHologramLineClickEvent implements HologramLineClickEvent {
