+++ plugin/src/main/java/me/filoghost/holographicdisplays/plugin/api/current/SimpleHologramLinePickupEvent.java
@@ -5,7 +5,7 @@
  */
 package me.filoghost.holographicdisplays.plugin.api.current;
 
-import me.filoghost.holographicdisplays.api.hologram.line.HologramLinePickupEvent;
+import me.filoghost.holographicdisplays.api.beta.hologram.line.HologramLinePickupEvent;
 import org.bukkit.entity.Player;
 
 class SimpleHologramLinePickupEvent implements HologramLinePickupEvent {
