+++ plugin/src/main/java/me/filoghost/holographicdisplays/plugin/internal/placeholder/ImmutablePlaceholder.java
@@ -5,7 +5,7 @@
  */
 package me.filoghost.holographicdisplays.plugin.internal.placeholder;
 
-import me.filoghost.holographicdisplays.api.placeholder.GlobalPlaceholder;
+import me.filoghost.holographicdisplays.api.beta.placeholder.GlobalPlaceholder;
 
 public class ImmutablePlaceholder implements GlobalPlaceholder {
